refactor lighthouse
-remove use of lbrynet-daemon, instead use just the lbrynet dht node
and blob manager to provide the minimal functionality required

-break out use of lbrycrd from lbrynet LBRYcrdWallet, avoiding a
headache when it is removed from lbrynet

-refactor availability and costs return fields to be more meaningful:
availability is a peer count from the lighthouse servers perspective, a
gauge of relative availability. and cost is the fee plus the stream
size, if known, which the client then can use to estimate the cost
locally.

-don’t take three days to start up from scratch (now ~10 seconds), and
don’t make my laptop cpu break a sweat.